src/vision/recognizer.py
import mss
import numpy as np
import cv2
import pytesseract
from PIL import Image

class Recognizer:
    def __init__(self, monitor_number=1):
        self.sct = mss.mss()
        # Part of the screen to capture
        # For now, we capture the whole screen. This should be configured.
        self.monitor = self.sct.monitors[monitor_number]

        # Configure Tesseract path if it's not in your system's PATH
        # pytesseract.pytesseract.tesseract_cmd = r'/usr/bin/tesseract' # Example for Linux

    def capture_screen(self):
        """Captures the screen and returns it as an OpenCV-compatible BGR NumPy array."""
        sct_img = self.sct.grab(self.monitor)
        # Convert to a NumPy array
        img_rgba = np.array(sct_img)
        # Convert RGBA to BGR
        img_bgr = cv2.cvtColor(img_rgba, cv2.COLOR_RGBA2BGR)
        return img_bgr

    def _convert_cv2_to_pil(self, cv2_image):
        """Converts an OpenCV BGR image to a PIL Image."""
        # OpenCV images are BGR, PIL images are RGB
        rgb_image = cv2.cvtColor(cv2_image, cv2.COLOR_BGR2RGB)
        return Image.fromarray(rgb_image)

    def detect_upgrade_event(self, screen_capture):
        """
        Detects an upgrade event and extracts item_id and new_level.

        This function will analyze a portion of the screen to detect
        if an item upgrade has occurred. It will return the item_id
        and new_level if an upgrade is detected.
        """
        # --- Step 1: Define Regions of Interest (ROIs) ---
        # These coordinates are highly dependent on the game's UI.
        # They would need to be determined by analyzing screenshots of the game.
        # Example ROIs (these are placeholders and need to be refined):
        # upgrade_popup_roi = {'top': 200, 'left': 400, 'width': 500, 'height': 300}
        # item_id_roi = {'top': 250, 'left': 450, 'width': 200, 'height': 50}
        # new_level_roi = {'top': 300, 'left': 450, 'width': 100, 'height': 30}

        # For demonstration, let's assume we're looking at a specific small region
        # where an "upgrade complete" message might appear.
        # This will need to be made dynamic based on actual UI analysis.
        # example_roi = screen_capture[200:250, 500:700] # Y:200-250, X:500-700

        # --- Step 2: Use Template Matching / Feature Detection ---
        # This would involve loading pre-defined images of UI elements (templates)
        # and using cv2.matchTemplate to find them in the screen_capture.
        # For example, a template for an "Upgrade Complete" button or a specific icon.
        # If a match is found with high confidence, it suggests an upgrade event.
        #
        # For now, we'll simulate a detection for demonstration purposes.
        # In a real scenario, this would be a sophisticated image processing pipeline.
        # if self._is_upgrade_popup_present(screen_capture, upgrade_popup_template):
        #     # Further analyze specific areas for item_id and new_level

        # --- Step 3: Optical Character Recognition (OCR) ---
        # If text is present in the ROIs (e.g., item ID or level number),
        # OCR can be used to extract it. pytesseract works well with PIL Images.
        #
        # pil_image_roi = self._convert_cv2_to_pil(example_roi)
        # text_from_roi = pytesseract.image_to_string(pil_image_roi, config='--psm 7') # psm 7 for single text line
        #
        # print(f"OCR text from ROI: {text_from_roi.strip()}")

        # --- Step 4: Compare Before/After States (Requires capturing previous screen) ---
        # To truly detect an *event* (change), you'd compare the current screen
        # with the screen from a few moments ago. This requires storing the previous
        # screen capture or hashes of it.

        # Placeholder for detected values
        # For now, return None, None, None as this is a complex implementation
        # and requires actual game UI analysis.
        # The logic here will be expanded significantly based on actual game UI.

        print("Scanning for upgrade event... (detailed placeholder)")
        return None, None, None
