requirements.txt
# For screen capturing and pixel analysis
mss
numpy
Pillow
opencv-python
pytesseract

# For the companion app UI
Flask
requests